feat : 전략패턴 제거
