+++ src/main/java/lottery_auto/domain/LotteryMachine.java
@@ -1,18 +1,21 @@
 package lottery_auto.domain;
 
-import lottery_auto.strategy.LotteryStrategy;
-import lottery_auto.strategy.LotteryStrategyImpl;
-
 import java.math.BigDecimal;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
+import static java.util.Collections.shuffle;
+import static java.util.stream.IntStream.rangeClosed;
+
 public final class LotteryMachine {
 
     private static final BigDecimal LOTTO_PRICE = new BigDecimal(1000);
-    private static final int COMPARISON_CRITERIA = 0;
-    private static final LotteryStrategy LOTTO_STRATEGY = new LotteryStrategyImpl();
+    private static final int LOTTO_IDX_ZERO = 0;
+    private static final int LOTTO_IDX = 6;
+    private static final int LOTTO_MIN_NUMBER = 1;
+    private static final int LOTTO_MAX_NUMBER = 45;
 
     private static LotteryMachine lotteryMachine = null;
 
@@ -34,7 +37,12 @@ public final class LotteryMachine {
     }
 
     public List<Integer> generateNumber(){
-        return LOTTO_STRATEGY.generate();
+        List<Integer> numbers = shuffleNumber(disposeNumber());
+
+        return new ArrayList<>(numbers.subList(LOTTO_IDX_ZERO, LOTTO_IDX))
+                .stream()
+                .sorted()
+                .collect(Collectors.toList());
     }
 
     public int purchase(final BigDecimal amount) {
@@ -45,7 +53,18 @@ public final class LotteryMachine {
     }
 
     public boolean validate(final BigDecimal amount){
-        return amount.compareTo(LOTTO_PRICE) >= COMPARISON_CRITERIA;
+        return amount.compareTo(LOTTO_PRICE) >= LOTTO_IDX_ZERO;
+    }
+
+    private List<Integer> disposeNumber() {
+        List<Integer> numbers = new ArrayList<>();
+        rangeClosed(LOTTO_MIN_NUMBER, LOTTO_MAX_NUMBER).forEach(numbers::add);
+        return numbers;
+    }
+
+    private List<Integer> shuffleNumber(List<Integer> numbers) {
+        shuffle(numbers);
+        return numbers;
     }
 
 }
