--- src/main/java/lottery_auto/strategy/LotteryStrategy.java
@@ -1,7 +1,0 @@
-package lottery_auto.strategy;
-
-import java.util.List;
-
-public interface LotteryStrategy {
-    List<Integer> generate();
-}
